+++ packages/shell/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [0.6.0] - 2026-06-07
+
+- Updated workspace package version alignment for the run-screen release.
+
 ## [0.5.0] - 2026-06-07
 
 - Updated workspace package version alignment for the local store + notifications release.